[XLA] Always be willing to duplicate widening kConvert instructions during fusion.
This has the effect of pushing widening kConvert HLOs into consumers.
This is what we want, because it means that the producer writes the
narrower type (e.g. f16) and the consumer reads it and internally
upcasts to the wider type (e.g. f32).  This lets the producer and
consumer both run faster, because they have to touch less memory.

PiperOrigin-RevId: 195546910