Use "```" (backtick) for code blocks in adding_an_op.md (#19187)
* Use "```" (backtick) for code blocks in adding_an_op.md

In adding_an_op.md, most of the code blocks uses "```" (backtick)
and annotations are added automatically. Though there was one
place where the code block are done with manual html code. This
is really error-prune and hard to change if there is an update
in the future.

This fix converts to "```c++" (backticks) so that it is easy
to maintain in the future.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Fix extra `\` at the beginning of the block

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>

* Update adding_an_op.md

add new lines where the <br/> tags were.